GetReturnsEvaluation( float[] returns ) returns a float now (it returned a double in previous version)

+++ QuantProject/b4_Business/a2_Strategies/EquityEvaluation/IEquityEvaluator.cs
@@ -29,6 +29,6 @@ namespace QuantProject.Business.Strategies.EquityEvaluation
 	/// </summary>
 	public interface IEquityEvaluator
 	{
-		double GetReturnsEvaluation( double[] returns );
+		float GetReturnsEvaluation( float[] returns );
 	}
 }
